Added README for Smith-Waterman example.
Also made some other cosmetic changes to the example's code, including
removing some inline keywords that seemed to be triggering a bug in gcc.
